Final test.

+++ 5_downscaling_parallel.py
@@ -26,8 +26,7 @@ if clean_previous_output and os.path.exists(general_output_folder): shutil.rmtre
 
 # make log folder and initialize logging
 log_file_folder = general_output_folder + "/global/log/"
-if clean_previous_output: 
-    shutil.rmtree(log_file_folder)
+if clean_previous_output and os.path.exists(log_file_folder): shutil.rmtree(log_file_folder)
 if os.path.exists(log_file_folder) == False: os.makedirs(log_file_folder)
 vos.initialize_logging(log_file_folder)
